refactor: split 4144-line visualize.py into a visualize/ package
The monolithic visualize.py is replaced by a package with one file per
visualization domain. Every public and private symbol is re-exported from
visualize/__init__.py, so existing imports (ludwig.cli, examples, tests) are
unaffected.

Package layout:
  visualize/_utils.py      — shared helpers: data loading, ground-truth processing,
                             validation, path generation
  visualize/training.py    — learning_curves / learning_curves_cli
  visualize/performance.py — compare_performance, compare_classifiers_* (18 functions)
  visualize/threshold.py   — confidence_thresholding_* + binary_threshold_vs_metric (14 functions)
  visualize/curves.py      — roc_curves, precision_recall_curves, calibration (12 functions)
  visualize/confusion.py   — confusion_matrix / confusion_matrix_cli
  visualize/hyperopt.py    — hyperopt_report, hyperopt_hiplot, hyperopt_results_to_dataframe
  visualize/cli.py         — get_visualizations_registry() + cli() entry point

Incidental fixes:
  - calibration_multiclass: two bare list references fixed to model_names_list
  - pyproject.toml: add RUF022 to ignore (thematic __all__ grouping preferred)
  - .pre-commit-config.yaml: exclude IMPROVEMENT_PLAN.md from blacken-docs/mdformat
