+++ .pre-commit-config.yaml
@@ -46,6 +46,7 @@ repos:
     hooks:
       - id: blacken-docs
         args: [--line-length=120]
+        exclude: "IMPROVEMENT_PLAN.md"
   - repo: https://github.com/hukkin/mdformat
     rev: 1.0.0
     hooks:
@@ -53,7 +54,7 @@ repos:
         additional_dependencies:
           - mdformat-gfm==1.0.0
           - mdformat_frontmatter==2.0.10
-        exclude: CHANGELOG.md
+        exclude: "CHANGELOG.md|IMPROVEMENT_PLAN.md"
   - repo: https://github.com/yoheimuta/protolint
     rev: v0.56.4
     hooks:
